+++ omnidesk/ui/main_window.py
@@ -95,10 +95,6 @@ class MainWindow(QMainWindow):
         self._refresh_action.setShortcut(QKeySequence(Qt.Key.Key_F5))
         self._refresh_action.triggered.connect(self._handle_refresh)
 
-        self._go_up_action = QAction("Go Up", self)
-        self._go_up_action.setShortcut(QKeySequence(Qt.Key.Key_Backspace))
-        self._go_up_action.triggered.connect(self._handle_go_up)
-
         self._toggle_view_action = QAction("Switch to Column View", self)
         self._toggle_view_action.setShortcut(QKeySequence("Ctrl+Shift+C"))
         self._toggle_view_action.triggered.connect(self._handle_toggle_view)
@@ -116,7 +112,6 @@ class MainWindow(QMainWindow):
             self._close_tab_action,
             self._open_folder_action,
             self._refresh_action,
-            self._go_up_action,
             self._toggle_view_action,
             self._next_tab_action,
             self._previous_tab_action,
@@ -129,7 +124,6 @@ class MainWindow(QMainWindow):
         toolbar.addAction(self._new_tab_action)
         toolbar.addSeparator()
         toolbar.addAction(self._open_folder_action)
-        toolbar.addAction(self._go_up_action)
         toolbar.addAction(self._refresh_action)
         toolbar.addSeparator()
         toolbar.addAction(self._toggle_view_action)